Roughed in Shell for SQLite Data Support (This will be the default option for regular users) Lots of updates to Data for changes in model.  To that end cleaning up the Repository Breakdown.  Lines are pretty blurred. Started moving ExternalIntegration Service to more of a service orientation pattern, but some of the common models fell short, probable need to re-think or at least push up things like the Macro Runtime Settings (which are emulator specific) Some minor work to the client Added dependency to SQLite and Entity Framework.  It's notable that the Entity Framework dependency is not used, it's just brought in with the official nuget package.

+++ src/VF.Macros.Common/Models/Macro/Action.cs
@@ -38,7 +38,7 @@ namespace VF.Macros.Common.Models.Macro
         /// <summary>
         /// The Action Delay
         /// </summary>
-        public long ActionDelay { get; set; }
+        public int ActionDelay { get; set; }
 
         /// <summary>
         /// Initialize Action Data Contract
